+++ src/main/java/net/frozenblock/wilderwild/mixin/datafix/DataFixersMixin.java
@@ -84,10 +84,8 @@ public class DataFixersMixin {
 		builder.addFixer(new AddNewChoices(schema, WWConstants.string("butterfly"), References.ENTITY));
 		builder.addFixer(new AddNewChoices(schema, WWConstants.string("tumbleweed"), References.ENTITY));
 		builder.addFixer(new AddNewChoices(schema, WWConstants.string("coconut"), References.ENTITY));
-		builder.addFixer(new AddNewChoices(schema, WWConstants.string("sculk_spreader"), References.ENTITY));
 		builder.addFixer(new AddNewChoices(schema, WWConstants.string("scorched"), References.ENTITY));
 		builder.addFixer(new AddNewChoices(schema, WWConstants.string("moobloom"), References.ENTITY));
-		builder.addFixer(new AddNewChoices(schema, WWConstants.string("falling_leaves"), References.ENTITY));
 		return schema;
 	}