+++ CHANGELOG.md
@@ -67,6 +67,7 @@ hi
 - Added a config option to toggle pools of water generating in River biomes.
   - This is now disabled by default, unlike before.
 - Scorched now set entities on fire upon attacking.
+- Removed the `wilderwild:sculk_spreader` entity, and updated the `spreadsculk` command to place all Sculk at once.
 
 # Music
 - Added three new tracks by Willow/pictochats_.